+++ src/mosaic.rs
@@ -37,6 +37,7 @@ use crate::mosaic::twos::build_2_mosaic;
 mod twos;
 mod threes;
 mod fours;
+mod testutils;
 
 const SPACING_SIZE: u32 = 10;
 const MAX_SIZE: u32 = 4000;
@@ -346,380 +347,19 @@ fn build_mosaic<const LEN: usize>(mosaic: MosaicImageDims<LEN>, images: [RgbImag
 
 #[cfg(test)]
 mod tests {
-    use std::fs;
-
-    use image::{Rgb, RgbImage};
-
     use crate::mosaic;
-
-    const BLACK: Rgb<u8> = Rgb([0, 0, 0]);
-    const RED: Rgb<u8> = Rgb([255, 0, 0]);
-    const BLUE: Rgb<u8> = Rgb([0, 0, 255]);
-    const GREEN: Rgb<u8> = Rgb([0, 255, 0]);
-    const PURPLE: Rgb<u8> = Rgb([255, 64, 255]);
-    const TEST_RESULT_DIR: &str = "./mosaic_tests/";
-
-    fn create_with_colour(width: u32, height: u32, colour: Rgb<u8>) -> RgbImage {
-        let mut img = RgbImage::new(width, height);
-
-        for x in 0..width {
-            for y in 0..height {
-                img.put_pixel(x, y, colour);
-                img.put_pixel(x, y, colour);
-            }
-        }
-
-        img
-    }
-
-    fn is_colour_at_pixel(x: u32, y: u32, image: &RgbImage, colour: Rgb<u8>) -> bool {
-        image.get_pixel(x, y).eq(&colour)
-    }
-
-    fn is_colour_in_range(start_x: u32, start_y: u32, end_x: u32, end_y: u32, image: &RgbImage, colour: Rgb<u8>) -> bool {
-        for x in start_x..end_x {
-            for y in start_y..end_y {
-                if !is_colour_at_pixel(x, y, image, colour) {
-                    return false;
-                }
-            }
-        }
-        true
-    }
-
-    fn has_black_vertical_line(x: u32, image: &RgbImage) -> bool {
-        is_colour_in_range(x, 0, x, image.height(), image, BLACK)
-    }
-
-    fn has_black_horizontal_line(y: u32, image: &RgbImage) -> bool {
-        is_colour_in_range(0, y, image.width(), y, image, BLACK)
-    }
-
-    fn has_black_vertical_line_partial(x: u32, start_y: u32, end_y: u32, image: &RgbImage) -> bool {
-        is_colour_in_range(x, start_y, x, end_y, image, BLACK)
-    }
-
-    fn has_black_horizontal_line_partial(y: u32, start_x: u32, end_x: u32, image: &RgbImage) -> bool {
-        is_colour_in_range(start_x, y, end_x, y, image, BLACK)
-    }
-
-    fn save_result(result: &RgbImage, filename: &str) {
-        let file_path = [TEST_RESULT_DIR, filename, ".png"].join("");
-        fs::create_dir_all(TEST_RESULT_DIR).unwrap();
-        result.save(file_path).unwrap();
-    }
-
-    #[test]
-    fn mosaic_2_left_right() {
-        let left = create_with_colour(100, 400, RED);
-        let right = create_with_colour(200, 400, BLUE);
-
-        let result = mosaic(vec![left, right]);
-
-        save_result(&result, "2-left_right");
-        assert!(is_colour_in_range(0, 0, 100, 400, &result, RED));
-        assert!(is_colour_in_range(120, 0, 300, 400, &result, BLUE));
-        assert!(has_black_vertical_line(105, &result));
-    }
-
-    #[test]
-    fn mosaic_2_top_bottom() {
-        let top = create_with_colour(400, 200, RED);
-        let bottom = create_with_colour(400, 100, BLUE);
-
-        let result = mosaic(vec![top, bottom]);
-
-        save_result(&result, "2-top_bottom");
-        assert!(is_colour_in_range(0, 0, 400, 200, &result, RED));
-        assert!(is_colour_in_range(0, 220, 400, 300, &result, BLUE));
-        assert!(has_black_horizontal_line(205, &result));
-    }
-
-    #[test]
-    fn mosaic_3_three_cols() {
-        let left = create_with_colour(100, 400, RED);
-        let mid = create_with_colour(200, 400, BLUE);
-        let right = create_with_colour(100, 400, GREEN);
-
-        let result = mosaic(vec![left, mid, right]);
-
-        save_result(&result, "3-three_cols");
-        assert!(is_colour_in_range(0, 0, 100, 400, &result, RED));
-        assert!(has_black_vertical_line(105, &result));
-        assert!(is_colour_in_range(120, 0, 300, 400, &result, BLUE));
-        assert!(has_black_vertical_line(315, &result));
-        assert!(is_colour_in_range(330, 0, 400, 400, &result, GREEN));
-    }
-
-    #[test]
-    fn mosaic_3_top_top_bottom() {
-        let top_left = create_with_colour(200, 300, RED);
-        let top_right = create_with_colour(200, 300, BLUE);
-        let bottom = create_with_colour(400, 100, GREEN);
-
-        let result = mosaic(vec![top_left, top_right, bottom]);
-
-        save_result(&result, "3-top_top_bottom");
-        assert!(is_colour_in_range(0, 0, 200, 300, &result, RED));
-        assert!(has_black_vertical_line_partial(205, 0, 300, &result));
-        assert!(is_colour_in_range(220, 0, 400, 300, &result, BLUE));
-        assert!(has_black_horizontal_line(305, &result));
-        assert!(is_colour_in_range(0, 320, 400, 400, &result, GREEN));
-    }
-
-    #[test]
-    fn mosaic_3_left_left_right() {
-        let left_top = create_with_colour(300, 200, RED);
-        let left_bot = create_with_colour(300, 200, BLUE);
-        let right = create_with_colour(100, 400, GREEN);
-
-        let result = mosaic(vec![left_top, left_bot, right]);
-
-        save_result(&result, "3-left_left_right");
-        assert!(is_colour_in_range(0, 0, 300, 200, &result, RED));
-        assert!(has_black_horizontal_line_partial(205, 0, 300, &result));
-        assert!(is_colour_in_range(0, 220, 300, 400, &result, BLUE));
-        assert!(has_black_vertical_line(305, &result));
-        assert!(is_colour_in_range(320, 0, 400, 400, &result, GREEN));
-    }
-
-    #[test]
-    fn mosaic_3_left_right_right() {
-        let left = create_with_colour(100, 400, RED);
-        let right_top = create_with_colour(300, 200, BLUE);
-        let right_bot = create_with_colour(300, 200, GREEN);
-
-        let result = mosaic(vec![left, right_top, right_bot]);
-
-        save_result(&result, "3-left_right_right");
-        assert!(is_colour_in_range(0, 0, 100, 400, &result, RED));
-        assert!(has_black_vertical_line(105, &result));
-        assert!(is_colour_in_range(120, 0, 400, 200, &result, BLUE));
-        assert!(has_black_horizontal_line_partial(205, 120, 400, &result));
-        assert!(is_colour_in_range(120, 220, 400, 400, &result, GREEN));
-    }
-
-    #[test]
-    fn mosaic_3_top_bottom_bottom() {
-        let top = create_with_colour(400, 100, RED);
-        let bot_left = create_with_colour(200, 300, BLUE);
-        let bot_right = create_with_colour(200, 300, GREEN);
-
-        let result = mosaic(vec![top, bot_left, bot_right]);
-
-        save_result(&result, "3-top_bottom_bottom");
-        assert!(is_colour_in_range(0, 0, 400, 100, &result, RED));
-        assert!(has_black_horizontal_line(105, &result));
-        assert!(is_colour_in_range(0, 120, 200, 400, &result, BLUE));
-        assert!(has_black_vertical_line_partial(205, 120, 400, &result));
-        assert!(is_colour_in_range(220, 120, 400, 400, &result, GREEN));
-    }
-
-    #[test]
-    fn mosaic_3_three_rows() {
-        let row1 = create_with_colour(300, 100, RED);
-        let row2 = create_with_colour(300, 100, BLUE);
-        let row3 = create_with_colour(300, 100, GREEN);
-
-        let result = mosaic(vec![row1, row2, row3]);
-
-        save_result(&result, "3-three_rows");
-        assert!(is_colour_in_range(0, 0, 300, 100, &result, RED));
-        assert!(has_black_horizontal_line(105, &result));
-        assert!(is_colour_in_range(0, 120, 300, 200, &result, BLUE));
-        assert!(has_black_horizontal_line(215, &result));
-        assert!(is_colour_in_range(0, 230, 300, 300, &result, GREEN));
-    }
-
-    #[test]
-    fn mosaic_4_four_cols() {
-        let col1 = create_with_colour(100, 400, RED);
-        let col2 = create_with_colour(100, 400, BLUE);
-        let col3 = create_with_colour(100, 400, GREEN);
-        let col4 = create_with_colour(100, 400, PURPLE);
-
-        let result = mosaic(vec![col1, col2, col3, col4]);
-
-        save_result(&result, "4-four_cols");
-        assert!(is_colour_in_range(0, 0, 100, 400, &result, RED));
-        assert!(has_black_vertical_line(105, &result));
-        assert!(is_colour_in_range(120, 0, 200, 400, &result, BLUE));
-        assert!(has_black_vertical_line(215, &result));
-        assert!(is_colour_in_range(230, 0, 300, 400, &result, GREEN));
-        assert!(has_black_vertical_line(325, &result));
-        assert!(is_colour_in_range(340, 0, 400, 400, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_four_rows() {
-        let row1 = create_with_colour(400, 100, RED);
-        let row2 = create_with_colour(400, 100, BLUE);
-        let row3 = create_with_colour(400, 100, GREEN);
-        let row4 = create_with_colour(400, 100, PURPLE);
-
-        let result = mosaic(vec![row1, row2, row3, row4]);
-
-        save_result(&result, "4-four_rows");
-        assert!(is_colour_in_range(0, 0, 400, 100, &result, RED));
-        assert!(has_black_horizontal_line(105, &result));
-        assert!(is_colour_in_range(0, 120, 400, 200, &result, BLUE));
-        assert!(has_black_horizontal_line(215, &result));
-        assert!(is_colour_in_range(0, 230, 400, 300, &result, GREEN));
-        assert!(has_black_horizontal_line(325, &result));
-        assert!(is_colour_in_range(0, 340, 400, 400, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_two_rows_of_two() {
-        let top_left = create_with_colour(100, 200, RED);
-        let top_right = create_with_colour(300, 200, BLUE);
-        let bot_left = create_with_colour(300, 200, GREEN);
-        let bot_right = create_with_colour(100, 200, PURPLE);
-
-        let result = mosaic(vec![top_left, top_right, bot_left, bot_right]);
-
-        save_result(&result, "4-two_rows_of_two");
-        assert!(is_colour_in_range(0, 0, 100, 200, &result, RED));
-        assert!(has_black_vertical_line_partial(105, 0, 200, &result));
-        assert!(is_colour_in_range(120, 0, 400, 200, &result, BLUE));
-        assert!(has_black_horizontal_line(205, &result));
-        assert!(is_colour_in_range(0, 220, 300, 400, &result, GREEN));
-        assert!(has_black_vertical_line_partial(305, 220, 400, &result));
-        assert!(is_colour_in_range(320, 220, 400, 400, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_two_rows_one_three() {
-        let top = create_with_colour(300, 200, RED);
-        let bot_left = create_with_colour(100, 100, BLUE);
-        let bot_mid = create_with_colour(100, 100, GREEN);
-        let bot_right = create_with_colour(100, 100, PURPLE);
-
-        let result = mosaic(vec![top, bot_left, bot_mid, bot_right]);
-
-        save_result(&result, "4-two_rows_one_three");
-        assert!(is_colour_in_range(0, 0, 300, 200, &result, RED));
-        assert!(has_black_horizontal_line(205, &result));
-        assert!(is_colour_in_range(0, 230, 100, 300, &result, BLUE));
-        assert!(has_black_vertical_line_partial(105, 220, 300, &result));
-        assert!(is_colour_in_range(120, 230, 200, 300, &result, GREEN));
-        assert!(has_black_vertical_line_partial(215, 220, 300, &result));
-        assert!(is_colour_in_range(230, 230, 300, 300, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_two_rows_three_one() {
-        let top_left = create_with_colour(100, 100, RED);
-        let top_mid = create_with_colour(100, 100, BLUE);
-        let top_right = create_with_colour(100, 100, GREEN);
-        let bottom = create_with_colour(300, 200, PURPLE);
-
-        let result = mosaic(vec![top_left, top_mid, top_right, bottom]);
-
-        save_result(&result, "4-two_rows_three_one");
-        assert!(is_colour_in_range(0, 0, 100, 100, &result, RED));
-        assert!(has_black_vertical_line_partial(105, 0, 100, &result));
-        assert!(is_colour_in_range(120, 0, 200, 100, &result, BLUE));
-        assert!(has_black_vertical_line_partial(215, 0, 100, &result));
-        assert!(is_colour_in_range(230, 0, 300, 100, &result, GREEN));
-        assert!(has_black_horizontal_line(105, &result));
-        assert!(is_colour_in_range(0, 120, 300, 300, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_two_columns_one_three() {
-        let left = create_with_colour(200, 300, RED);
-        let right_top = create_with_colour(100, 100, BLUE);
-        let right_mid = create_with_colour(100, 100, GREEN);
-        let right_bot = create_with_colour(100, 100, PURPLE);
-
-        let result = mosaic(vec![left, right_top, right_mid, right_bot]);
-
-        save_result(&result, "4-two_columns_one_three");
-        assert!(is_colour_in_range(0, 0, 200, 300, &result, RED));
-        assert!(has_black_vertical_line(105, &result));
-        assert!(is_colour_in_range(230, 0, 300, 100, &result, BLUE));
-        assert!(has_black_horizontal_line_partial(105, 220, 300, &result));
-        assert!(is_colour_in_range(230, 120, 300, 200, &result, GREEN));
-        assert!(has_black_horizontal_line_partial(215, 220, 300, &result));
-        assert!(is_colour_in_range(230, 230, 300, 300, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_two_columns_three_one() {
-        let left_top = create_with_colour(100, 100, RED);
-        let left_mid = create_with_colour(100, 100, BLUE);
-        let left_bot = create_with_colour(100, 100, GREEN);
-        let right = create_with_colour(200, 300, PURPLE);
-
-        let result = mosaic(vec![left_top, left_mid, left_bot, right]);
-
-        save_result(&result, "4-two_columns_three_one");
-        assert!(is_colour_in_range(0, 0, 100, 100, &result, RED));
-        assert!(has_black_horizontal_line_partial(105, 0, 200, &result));
-        assert!(is_colour_in_range(0, 120, 100, 200, &result, BLUE));
-        assert!(has_black_horizontal_line_partial(215, 0, 200, &result));
-        assert!(is_colour_in_range(0, 230, 100, 300, &result, GREEN));
-        assert!(has_black_vertical_line(105, &result));
-        assert!(is_colour_in_range(220, 230, 300, 300, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_three_rows_211() {
-        let top_left = create_with_colour(300, 200, RED);
-        let top_right = create_with_colour(300, 200, BLUE);
-        let mid = create_with_colour(600, 200, GREEN);
-        let bot = create_with_colour(600, 200, PURPLE);
-
-        let result = mosaic(vec![top_left, top_right, mid, bot]);
-
-        save_result(&result, "4-three_rows_211");
-        assert!(is_colour_in_range(0, 0, 300, 200, &result, RED));
-        assert!(has_black_vertical_line_partial(305, 0, 200, &result));
-        assert!(is_colour_in_range(320, 0, 600, 200, &result, BLUE));
-        assert!(has_black_horizontal_line(205, &result));
-        assert!(is_colour_in_range(0, 220, 600, 400, &result, GREEN));
-        assert!(has_black_horizontal_line(415, &result));
-        assert!(is_colour_in_range(0, 430, 600, 600, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_three_rows_121() {
-        let top = create_with_colour(600, 200, RED);
-        let mid_left = create_with_colour(300, 200, BLUE);
-        let mid_right = create_with_colour(300, 200, GREEN);
-        let bot = create_with_colour(600, 200, PURPLE);
-
-        let result = mosaic(vec![top, mid_left, mid_right, bot]);
-
-        save_result(&result, "4-three_rows_121");
-        assert!(is_colour_in_range(0, 0, 600, 200, &result, RED));
-        assert!(has_black_horizontal_line(205, &result));
-        assert!(is_colour_in_range(0, 220, 300, 400, &result, BLUE));
-        assert!(has_black_vertical_line_partial(305, 210, 400, &result));
-        assert!(is_colour_in_range(320, 220, 600, 400, &result, GREEN));
-        assert!(has_black_horizontal_line(415, &result));
-        assert!(is_colour_in_range(0, 430, 600, 600, &result, PURPLE));
-    }
-
-    #[test]
-    fn mosaic_4_three_rows_112() {
-        let top = create_with_colour(600, 200, RED);
-        let mid = create_with_colour(600, 200, BLUE);
-        let bot_left = create_with_colour(300, 200, GREEN);
-        let bot_right = create_with_colour(300, 200, PURPLE);
-
-        let result = mosaic(vec![top, mid, bot_left, bot_right]);
-
-        save_result(&result, "4-three_rows_112");
-        assert!(is_colour_in_range(0, 0, 600, 200, &result, RED));
-        assert!(has_black_horizontal_line(205, &result));
-        assert!(is_colour_in_range(0, 220, 600, 400, &result, BLUE));
-        assert!(has_black_horizontal_line(415, &result));
-        assert!(is_colour_in_range(0, 430, 300, 600, &result, GREEN));
-        assert!(has_black_vertical_line_partial(305, 420, 600, &result));
-        assert!(is_colour_in_range(320, 430, 600, 600, &result, PURPLE));
-    }
+    use crate::mosaic::testutils::{
+        BLUE,
+        create_with_colour,
+        GREEN,
+        has_black_horizontal_line,
+        has_black_vertical_line,
+        has_black_vertical_line_partial,
+        is_colour_in_range,
+        PURPLE,
+        RED,
+        save_result,
+    };
 
     #[test]
     fn pick_less_square_option_for_better_scaling_ratio() {
