+++ src/mosaic/fours.rs
@@ -428,3 +428,211 @@ fn three_columns_112_4_mosaic(first: Size, second: Size, third: Size, fourth: Si
         ],
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::mosaic;
+    use crate::mosaic::testutils::{
+        BLUE,
+        create_with_colour,
+        GREEN,
+        has_black_horizontal_line,
+        has_black_horizontal_line_partial,
+        has_black_vertical_line,
+        has_black_vertical_line_partial,
+        is_colour_in_range,
+        PURPLE,
+        RED,
+        save_result,
+    };
+
+    #[test]
+    fn mosaic_4_four_cols() {
+        let col1 = create_with_colour(100, 400, RED);
+        let col2 = create_with_colour(100, 400, BLUE);
+        let col3 = create_with_colour(100, 400, GREEN);
+        let col4 = create_with_colour(100, 400, PURPLE);
+
+        let result = mosaic(vec![col1, col2, col3, col4]);
+
+        save_result(&result, "4-four_cols");
+        assert!(is_colour_in_range(0, 0, 100, 400, &result, RED));
+        assert!(has_black_vertical_line(105, &result));
+        assert!(is_colour_in_range(120, 0, 200, 400, &result, BLUE));
+        assert!(has_black_vertical_line(215, &result));
+        assert!(is_colour_in_range(230, 0, 300, 400, &result, GREEN));
+        assert!(has_black_vertical_line(325, &result));
+        assert!(is_colour_in_range(340, 0, 400, 400, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_four_rows() {
+        let row1 = create_with_colour(400, 100, RED);
+        let row2 = create_with_colour(400, 100, BLUE);
+        let row3 = create_with_colour(400, 100, GREEN);
+        let row4 = create_with_colour(400, 100, PURPLE);
+
+        let result = mosaic(vec![row1, row2, row3, row4]);
+
+        save_result(&result, "4-four_rows");
+        assert!(is_colour_in_range(0, 0, 400, 100, &result, RED));
+        assert!(has_black_horizontal_line(105, &result));
+        assert!(is_colour_in_range(0, 120, 400, 200, &result, BLUE));
+        assert!(has_black_horizontal_line(215, &result));
+        assert!(is_colour_in_range(0, 230, 400, 300, &result, GREEN));
+        assert!(has_black_horizontal_line(325, &result));
+        assert!(is_colour_in_range(0, 340, 400, 400, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_two_rows_of_two() {
+        let top_left = create_with_colour(100, 200, RED);
+        let top_right = create_with_colour(300, 200, BLUE);
+        let bot_left = create_with_colour(300, 200, GREEN);
+        let bot_right = create_with_colour(100, 200, PURPLE);
+
+        let result = mosaic(vec![top_left, top_right, bot_left, bot_right]);
+
+        save_result(&result, "4-two_rows_of_two");
+        assert!(is_colour_in_range(0, 0, 100, 200, &result, RED));
+        assert!(has_black_vertical_line_partial(105, 0, 200, &result));
+        assert!(is_colour_in_range(120, 0, 400, 200, &result, BLUE));
+        assert!(has_black_horizontal_line(205, &result));
+        assert!(is_colour_in_range(0, 220, 300, 400, &result, GREEN));
+        assert!(has_black_vertical_line_partial(305, 220, 400, &result));
+        assert!(is_colour_in_range(320, 220, 400, 400, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_two_rows_one_three() {
+        let top = create_with_colour(300, 200, RED);
+        let bot_left = create_with_colour(100, 100, BLUE);
+        let bot_mid = create_with_colour(100, 100, GREEN);
+        let bot_right = create_with_colour(100, 100, PURPLE);
+
+        let result = mosaic(vec![top, bot_left, bot_mid, bot_right]);
+
+        save_result(&result, "4-two_rows_one_three");
+        assert!(is_colour_in_range(0, 0, 300, 200, &result, RED));
+        assert!(has_black_horizontal_line(205, &result));
+        assert!(is_colour_in_range(0, 230, 100, 300, &result, BLUE));
+        assert!(has_black_vertical_line_partial(105, 220, 300, &result));
+        assert!(is_colour_in_range(120, 230, 200, 300, &result, GREEN));
+        assert!(has_black_vertical_line_partial(215, 220, 300, &result));
+        assert!(is_colour_in_range(230, 230, 300, 300, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_two_rows_three_one() {
+        let top_left = create_with_colour(100, 100, RED);
+        let top_mid = create_with_colour(100, 100, BLUE);
+        let top_right = create_with_colour(100, 100, GREEN);
+        let bottom = create_with_colour(300, 200, PURPLE);
+
+        let result = mosaic(vec![top_left, top_mid, top_right, bottom]);
+
+        save_result(&result, "4-two_rows_three_one");
+        assert!(is_colour_in_range(0, 0, 100, 100, &result, RED));
+        assert!(has_black_vertical_line_partial(105, 0, 100, &result));
+        assert!(is_colour_in_range(120, 0, 200, 100, &result, BLUE));
+        assert!(has_black_vertical_line_partial(215, 0, 100, &result));
+        assert!(is_colour_in_range(230, 0, 300, 100, &result, GREEN));
+        assert!(has_black_horizontal_line(105, &result));
+        assert!(is_colour_in_range(0, 120, 300, 300, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_two_columns_one_three() {
+        let left = create_with_colour(200, 300, RED);
+        let right_top = create_with_colour(100, 100, BLUE);
+        let right_mid = create_with_colour(100, 100, GREEN);
+        let right_bot = create_with_colour(100, 100, PURPLE);
+
+        let result = mosaic(vec![left, right_top, right_mid, right_bot]);
+
+        save_result(&result, "4-two_columns_one_three");
+        assert!(is_colour_in_range(0, 0, 200, 300, &result, RED));
+        assert!(has_black_vertical_line(105, &result));
+        assert!(is_colour_in_range(230, 0, 300, 100, &result, BLUE));
+        assert!(has_black_horizontal_line_partial(105, 220, 300, &result));
+        assert!(is_colour_in_range(230, 120, 300, 200, &result, GREEN));
+        assert!(has_black_horizontal_line_partial(215, 220, 300, &result));
+        assert!(is_colour_in_range(230, 230, 300, 300, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_two_columns_three_one() {
+        let left_top = create_with_colour(100, 100, RED);
+        let left_mid = create_with_colour(100, 100, BLUE);
+        let left_bot = create_with_colour(100, 100, GREEN);
+        let right = create_with_colour(200, 300, PURPLE);
+
+        let result = mosaic(vec![left_top, left_mid, left_bot, right]);
+
+        save_result(&result, "4-two_columns_three_one");
+        assert!(is_colour_in_range(0, 0, 100, 100, &result, RED));
+        assert!(has_black_horizontal_line_partial(105, 0, 200, &result));
+        assert!(is_colour_in_range(0, 120, 100, 200, &result, BLUE));
+        assert!(has_black_horizontal_line_partial(215, 0, 200, &result));
+        assert!(is_colour_in_range(0, 230, 100, 300, &result, GREEN));
+        assert!(has_black_vertical_line(105, &result));
+        assert!(is_colour_in_range(220, 230, 300, 300, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_three_rows_211() {
+        let top_left = create_with_colour(300, 200, RED);
+        let top_right = create_with_colour(300, 200, BLUE);
+        let mid = create_with_colour(600, 200, GREEN);
+        let bot = create_with_colour(600, 200, PURPLE);
+
+        let result = mosaic(vec![top_left, top_right, mid, bot]);
+
+        save_result(&result, "4-three_rows_211");
+        assert!(is_colour_in_range(0, 0, 300, 200, &result, RED));
+        assert!(has_black_vertical_line_partial(305, 0, 200, &result));
+        assert!(is_colour_in_range(320, 0, 600, 200, &result, BLUE));
+        assert!(has_black_horizontal_line(205, &result));
+        assert!(is_colour_in_range(0, 220, 600, 400, &result, GREEN));
+        assert!(has_black_horizontal_line(415, &result));
+        assert!(is_colour_in_range(0, 430, 600, 600, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_three_rows_121() {
+        let top = create_with_colour(600, 200, RED);
+        let mid_left = create_with_colour(300, 200, BLUE);
+        let mid_right = create_with_colour(300, 200, GREEN);
+        let bot = create_with_colour(600, 200, PURPLE);
+
+        let result = mosaic(vec![top, mid_left, mid_right, bot]);
+
+        save_result(&result, "4-three_rows_121");
+        assert!(is_colour_in_range(0, 0, 600, 200, &result, RED));
+        assert!(has_black_horizontal_line(205, &result));
+        assert!(is_colour_in_range(0, 220, 300, 400, &result, BLUE));
+        assert!(has_black_vertical_line_partial(305, 210, 400, &result));
+        assert!(is_colour_in_range(320, 220, 600, 400, &result, GREEN));
+        assert!(has_black_horizontal_line(415, &result));
+        assert!(is_colour_in_range(0, 430, 600, 600, &result, PURPLE));
+    }
+
+    #[test]
+    fn mosaic_4_three_rows_112() {
+        let top = create_with_colour(600, 200, RED);
+        let mid = create_with_colour(600, 200, BLUE);
+        let bot_left = create_with_colour(300, 200, GREEN);
+        let bot_right = create_with_colour(300, 200, PURPLE);
+
+        let result = mosaic(vec![top, mid, bot_left, bot_right]);
+
+        save_result(&result, "4-three_rows_112");
+        assert!(is_colour_in_range(0, 0, 600, 200, &result, RED));
+        assert!(has_black_horizontal_line(205, &result));
+        assert!(is_colour_in_range(0, 220, 600, 400, &result, BLUE));
+        assert!(has_black_horizontal_line(415, &result));
+        assert!(is_colour_in_range(0, 430, 300, 600, &result, GREEN));
+        assert!(has_black_vertical_line_partial(305, 420, 600, &result));
+        assert!(is_colour_in_range(320, 430, 600, 600, &result, PURPLE));
+    }
+}
